Refactor egor solver files
