+++ crates/ego/src/solver/mod.rs
@@ -1,8 +1,9 @@
 mod egor_config;
-mod egor_impl;
 mod egor_service;
 mod egor_solver;
 mod egor_state;
+mod solver_computations;
+mod solver_impl;
 mod trego;
 
 pub use egor_config::*;
